+++ CHANGELOG.md
@@ -14,6 +14,9 @@ CHANGELOG
 - Fix Go SDK plugin acquisition for programs with vendored dependencies
   [#5286](https://github.com/pulumi/pulumi/pull/5286)
 
+- Python SDK: Add support for `Sequence[T]` for array types
+  [#5282](https://github.com/pulumi/pulumi/pull/5282)
+
 ## 2.9.2 (2020-08-31)
 
 - Alpha version of the Automation API for Go